refacto(obs.obs_codegen): rename some modules
make it clearer
