+++ obs/src/python/obs_codegen/__init__.py
@@ -1,4 +1,4 @@
 from . import c
-from .generator import BaseGenerator
-from .generator import LuaGenerator
-from .generator import HlslGenerator
+from .generators import BaseGenerator
+from .generators import LuaGenerator
+from .generators import HlslGenerator
