+++ obs/src/python/obs_codegen/colorcomponents.py
