+++ obs/src/python/obs_codegen/generators/_hlsl.py
@@ -5,10 +5,10 @@ import logging
 import colour
 
 from obs_codegen.c import HLSL_INDENT as INDENT
-from obs_codegen.generator import BaseGenerator
-from obs_codegen.entitities import Whitepoint
-from obs_codegen.entitities import Cat
-from obs_codegen.entitities import ColorspaceGamut
+from obs_codegen.generators import BaseGenerator
+from obs_codegen.colorcomponents import Whitepoint
+from obs_codegen.colorcomponents import Cat
+from obs_codegen.colorcomponents import ColorspaceGamut
 from obs_codegen.hlsl_utils import convert3x3MatrixToHlslStr
 from obs_codegen.hlsl_utils import generateCommentHeader
