+++ obs/src/scripts/build-colorspace_core.hlsl.py
@@ -3,11 +3,11 @@ from typing import Type
 
 import colour
 
-from obs_codegen.entitities import Whitepoint
-from obs_codegen.entitities import Cat
-from obs_codegen.entitities import AssemblyColorspace
-from obs_codegen.entitities import ColorspaceGamut
-from obs_codegen.entitities import TransferFunction
+from obs_codegen.colorcomponents import Whitepoint
+from obs_codegen.colorcomponents import Cat
+from obs_codegen.colorcomponents import AssemblyColorspace
+from obs_codegen.colorcomponents import ColorspaceGamut
+from obs_codegen.colorcomponents import TransferFunction
 from obs_codegen import BaseGenerator
 from obs_codegen import HlslGenerator
 from obs_codegen import LuaGenerator
